refactor: modal context API 적용 및 리팩토링
