+++ src/components/AddCard/index.js
@@ -1,32 +1,15 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext } from 'react';
 
 import Nav from '../mixin/Nav';
 import Card from '../../common/Card';
 import NewCardForm from './NewCardForm';
 import { MODAL } from '../../constants/constant';
 import { CardContext } from '../../data/context/CardContext';
+import { ModalContext } from '../../data/context/ModalContext';
 
 const AddCard = () => {
   const { cardInfo } = useContext(CardContext);
-
-  const [openModalContent, setOpenModalContent] = useState({
-    isModalOpen: false,
-    modalContent: '',
-  });
-
-  const onOpenModal = (modalContent) => {
-    setOpenModalContent({
-      isModalOpen: true,
-      modalContent,
-    });
-  };
-
-  const onCloseModal = () => {
-    setOpenModalContent({
-      isModalOpen: false,
-      modalContent: '',
-    });
-  };
+  const { onOpenModal } = useContext(ModalContext);
 
   return (
     <>
@@ -37,12 +20,7 @@ const AddCard = () => {
           onClickCard={() => onOpenModal(MODAL.CARD_COLOR)}
         />
       </div>
-      <NewCardForm
-        onOpenModal={onOpenModal}
-        onCloseModal={onCloseModal}
-        openModalContent={openModalContent}
-        setOpenModalContent={setOpenModalContent}
-      />
+      <NewCardForm />
     </>
   );
 };
